Added gene and snp marts as parameters for the pipeline. And now functions that use Marts are checking if this is a valid mart
Now "vargen_pipeline" and "vargen_custom" are taking a gene_mart and a snp_mart as parameters. This is useful if ensembl main server is down. The user can create a gene_mart from a mirror. eg: "gene_mart <- connect_to_gene_ensembl("useast") and use it as parameter.

Also changed the manhattan plot function to avoid the "require(ggbio)" which is a bad practice. See the comments in the function for more details.