Solving issue #1 ("convert_gtex_to_rsids" does not work for InDels.)
The function now adds +1 to the position so that ensembl can fetch the correct InDel.

Tested with both inversions and deletions from GTEx v7 and v8.

The only issue remaining is when there is a SNP at the same location than the InDel, then the function gets both (the InDel and the SNP). One solution would be to look at the alleles or use a lookup table.