Corrected bug with format output
Sometimes a gene appears several time in "omim_genes" (notably because of alternative haplotypes).

For example MAPT is both on "chr17" and "CHR_HSCHR17_2_CTG5".

This can cause an error when using "format_output", due to the difference in rows between the variants and the gene name (when the gene name is unique R duplicates it as many time as needed, but here we have 2 or 3 time the same gene name).